requirements.txt
appnope==0.1.4
asttokens==3.0.1
cmake==4.2.1
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.20
decorator==5.2.1
dune_alugrid==2.11.0.5
dune_common==2.11.0.5
dune_geometry==2.11.0.5
dune_grid==2.11.0.5
executing==2.2.1
fonttools==4.61.1
ipykernel==7.2.0
ipython==9.10.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.8.0
jupyter_core==5.9.1
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
nest-asyncio==1.6.0
ninja==1.13.0
numpy==2.4.2
packaging==26.0
parso==0.8.6
pexpect==4.9.0
pillow==12.1.1
platformdirs==4.9.2
portalocker==3.2.0
prompt_toolkit==3.0.52
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
pyzmq==27.1.0
scipy==1.17.0
setuptools==82.0.0
six==1.17.0
stack-data==0.6.3
tornado==6.5.4
traitlets==5.14.3
-e git+https://github.com/joshuadendy/Virtual-Element-Method.git@a9dbe2adadee0dae8cea00fd3caced25d4127474#egg=VEM
wcwidth==0.6.0
wheel==0.46.3
